Avoid leaving stale/confusing job remarks when job is replaced
When an existing job is replaced with a new job (because the new job has
the same settings as the existing job) the job remarks need to be cleaned.
Otherwise a stale "acceptable_for" remark leads to a job treated as
acceptable when no such comment is present on openQA yet.

Related ticket: https://progress.opensuse.org/issues/161267